player がアイテムを取得した時のみ音を鳴らす
